landscape: Move search system prefs into its own component
