+++ pkg/grid/src/preferences/SystemPreferences.tsx
@@ -1,8 +1,7 @@
-import React, { ChangeEvent, PropsWithChildren, useCallback, useEffect, useState } from 'react';
+import React, { PropsWithChildren, useCallback } from 'react';
 import { Link, Route, RouteComponentProps, Switch, useRouteMatch } from 'react-router-dom';
 import { ErrorBoundary } from 'react-error-boundary';
 import classNames from 'classnames';
-import fuzzy from 'fuzzy';
 import { NotificationPrefs } from './NotificationPrefs';
 import { AboutSystem } from './about-system/AboutSystem';
 import { InterfacePrefs } from './InterfacePrefs';
@@ -25,7 +24,7 @@ import PencilIcon from '../components/icons/PencilIcon';
 import ForwardSlashIcon from '../components/icons/ForwardSlashIcon';
 import { useSystemUpdate } from '../logic/useSystemUpdate';
 import { Bullet } from '../components/icons/Bullet';
-import MagnifyingGlassIcon from '../components/icons/MagnifyingGlassIcon';
+import SearchSystemPreferences from './SearchSystemPrefences';
 
 interface SystemPreferencesSectionProps {
   url: string;
@@ -52,49 +51,6 @@ function SystemPreferencesSection({
   );
 }
 
-const navOptions: { route: string; title: string; icon: React.ReactElement }[] = [
-  {
-    route: 'help',
-    title: 'Help and Support',
-    icon: <HelpIcon className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'interface',
-    title: 'Interface Settings',
-    icon: <Interface className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'notifications',
-    title: 'Notifications',
-    icon: <BellIcon className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'appearance',
-    title: 'Appearance',
-    icon: <PencilIcon className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'shortcuts',
-    title: 'Shortcuts',
-    icon: <ForwardSlashIcon className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'privacy',
-    title: 'Attention & Privacy',
-    icon: <BurstIcon className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'security',
-    title: 'Log Out...',
-    icon: <LogoutIcon className="w-4 h-4 text-gray-600" />
-  },
-  {
-    route: 'system-updates',
-    title: 'About System',
-    icon: <TlonIcon className="w-4 h-4 text-gray-600" />
-  }
-];
-
 export const SystemPreferences = (props: RouteComponentProps<{ submenu: string }>) => {
   const { match, history } = props;
   const subMatch = useRouteMatch<{ submenu: string; desk?: string }>(
@@ -105,21 +61,6 @@ export const SystemPreferences = (props: RouteComponentProps<{ submenu: string }
   const filteredCharges = Object.values(charges).filter((charge) => charge.desk !== window.desk);
   const isMobile = useMedia('(max-width: 639px)');
   const settingsPath = isMobile ? `${match.url}/:submenu` : '/';
-  const [searchInput, setSearchInput] = useState('');
-  const [matchingNavOptions, setMatchingNavOptions] = useState<string[]>([]);
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const input = e.target as HTMLInputElement;
-    const value = input.value.trim();
-
-    setSearchInput(value);
-  };
-
-  useEffect(() => {
-    const results = fuzzy.filter(searchInput, navOptions, { extract: (obj) => obj.title });
-    const matches = results.map((el) => el.string);
-    setMatchingNavOptions(matches);
-  }, [searchInput]);
 
   const matchSub = useCallback(
     (target: string, desk?: string) => {
@@ -151,39 +92,7 @@ export const SystemPreferences = (props: RouteComponentProps<{ submenu: string }
         <Route exact={isMobile} path={match.url}>
           <aside className="self-start flex-none w-full py-4 font-semibold text-black border-r-2 sm:w-auto min-w-60 sm:py-8 sm:text-gray-600 border-gray-50 bg-white">
             <nav className="px-2 sm:px-6 flex flex-col">
-              <label className="relative flex items-center">
-                <span className="sr-only">Search Prefences</span>
-                <span className="absolute h-8 w-8 text-gray-400 flex items-center pl-2 inset-y-1 left-0">
-                  <MagnifyingGlassIcon />
-                </span>
-                <input
-                  className="input bg-gray-50 pl-8 placeholder:font-semibold mb-5 h-10"
-                  placeholder="Search Preferences"
-                  value={searchInput}
-                  onChange={handleChange}
-                />
-              </label>
-              <div className="relative">
-                {matchingNavOptions.length > 0 && searchInput !== '' ? (
-                  <div className="absolute -top-3 flex flex-col bg-white space-y-2 rounded-2xl shadow-md w-full py-3">
-                    {matchingNavOptions.map((opt) => {
-                      const matchingNavOption = navOptions.find((navOpt) => navOpt.title === opt);
-                      if (matchingNavOption !== undefined) {
-                        return (
-                          <Link
-                            className="flex px-2 py-3 items-center space-x-2 hover:text-black hover:bg-gray-50"
-                            to={subUrl(matchingNavOption.route)}
-                          >
-                            {matchingNavOption.icon}
-                            <span className="text-gray-900">{matchingNavOption?.title}</span>
-                          </Link>
-                        );
-                      }
-                      return null;
-                    })}
-                  </div>
-                ) : null}
-              </div>
+              <SearchSystemPreferences subUrl={subUrl} />
               <span className="text-gray-400 font-semibold pt-1 pl-2 pb-3 text-sm">Landscape</span>
               <ul className="space-y-1">
                 <SystemPreferencesSection
